@tgregg 2022-01-25
Note: the performance regression test failed for an unrelated reason (`The performance regression detected when benchmark the ion-java from the new commit with the test data: testList.10n and parameters: write::{format:"JSON",type:"FILE",api:"STREAMING"}
The following aspects have regressions: {Heap usage=-0.0614610633338160}`), which is just noise. We occasionally get false positives with the regression detector. We should try to drive that down, but I'll take it over a false negative.